+++ src/main/java/kr/codesqaud/cafe/user/UserController.java
@@ -4,7 +4,8 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import kr.codesqaud.cafe.login.LoginRequestDto;
+import kr.codesqaud.cafe.user.dto.LoginRequestDto;
+import kr.codesqaud.cafe.user.dto.SignUpRequestDto;
 import kr.codesqaud.cafe.web.SessionConstant;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
